refactor:#4 유저 수정하기 구현(닉네임, 프로필)
