+++ src/main/java/com/trashheroesbe/feature/district/domain/entity/District.java
@@ -1,6 +1,5 @@
-package com.trashheroesbe.feature.district.domain;
+package com.trashheroesbe.feature.district.domain.entity;
 
-import com.trashheroesbe.global.entity.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
