+++ src/main/java/com/trashheroesbe/feature/district/infrastructure/DistrictRepository.java
@@ -1,6 +1,6 @@
 package com.trashheroesbe.feature.district.infrastructure;
 
-import com.trashheroesbe.feature.district.domain.District;
+import com.trashheroesbe.feature.district.domain.entity.District;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DistrictRepository extends JpaRepository<District, Long> {
+public interface DistrictRepository extends JpaRepository<District, String> {
 
     @Query("""
         SELECT d
