+++ src/main/java/com/trashheroesbe/feature/user/api/UserControllerApi.java
@@ -11,7 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 @Tag(name = "User", description = "유저 관련 API")
 public interface UserControllerApi {
 
-    @Operation(summary = "유저 수정하기(patch)", description = "유저의 정보를 수정합니다.(nickname, profileImg, district")
+    @Operation(summary = "유저 수정하기(patch)", description = "유저의 정보를 수정합니다.(nickname, profileImg")
     ApiResponse<Long> updateUser(
         @Parameter(description = "수정할 사용자 정보 JSON")
         UpdateUserRequest request,
